Added prod-only flag (-p, --prod)
As for the -d flag, this changes makes it possible to only check production dependencies (dependencies field in package.json).

This change assumes the following:

1. Include prod if
	- --prod is on
	- --dev is off
2. Include dev if
	- --dev is on
	- --prod is off and --dev is off

That way, we make sure that no options will still include everything and that specifing both --prod and --dev with include both as well. Specifing only one of the two will include the specified option.

Re #36