Provide helpful errors on JSON.parse exceptions
This commit uses `json-parse-helpfulerror` module to parse JSONs instead
of JSON.parse. The benefit of using it, is that when the parse fails and
throws, it provides clues to what in the JSON we tried to parse was
invalid. So it's easier to figure out, what's wrong and fix it. The
performance should pretty much be the same, since it first tries to use
JSON.parse and only if it throws, it analyzes the cause.

closestPackage.sync that we call when we try to find the closest
package,json, can throw exceptions, so I wrapped it in a try catch to
give better message. I also opened a PR
[there](https://github.com/hughsk/closest-package/pull/1), so they will
also throw more helpful errors (e.g. the new errors will include the
path to the package.json that caused the exception, so users will know
what to fix).