Replaced `closest-package` with `pkg-up`
The `closest-package` doesn't only find the closest package.json, it
also reads it and parses it. We don't need it to do that since we read
and parse it later ourselves. It also complicates the code, since we
need to wrap it with try and catch to deal with possible throws.

So I switched it with `pkg-up` which just finds the location of the
file.