fix: updates hosted-git-info to version 5.x
Version 4.x of hosted-git-info has a dependency on a broken version of the lru-cache module, which
causes the error "ReferenceError: AbortController is not defined". The lru-cache module has been
fixed, and hosted-git-info has been updated to use the fixed version, but that update is only on
version 5.x of hosted-git-info. This change updates to version 5.x of hosted-git-info to pull in
that fix, although 5.x is a breaking change, the only breaking aspect is that versions of node prior
to 12 LTS are no longer supported. This should not be an issue as npm-check-updates doesn't support
older versions of node.

fix #1073