chore(type): no functional change - adds explicit types to various yarn & doctor data/functions
The intent here is to be able to reason about the types being operated on by the 'doctor' just by reading the code at the various call sites. Also, we want to give the linter/typscript a greater chance of detecting problems, primarily because we appear to have a problem with reading VersionSpecs for yarn workspaces within the 'doctor'.